Theme generated in /theme applies to the whole page.
