+++ src/lib/misc/utils.ts
@@ -1,3 +1,5 @@
+import { DynamicScheme, MaterialDynamicColors } from "@material/material-color-utilities";
+
 export type Color =
   | "primary"
   | "onPrimary"
@@ -38,6 +40,60 @@ export type Color =
   | "surfaceTint";
 export type SerializedScheme = Record<Color, number>;
 
+export const pairs = [
+  ["primary", "onPrimary"],
+  ["primaryContainer", "onPrimaryContainer"],
+  ["secondary", "onSecondary"],
+  ["secondaryContainer", "onSecondaryContainer"],
+  ["tertiary", "onTertiary"],
+  ["tertiaryContainer", "onTertiaryContainer"],
+  ["background", "onBackground"],
+  ["surface", "onSurface"],
+  ["inverseSurface", "inverseOnSurface"],
+  ["surfaceVariant", "onSurfaceVariant"],
+  ["error", "onError"],
+  ["errorContainer", "onErrorContainer"],
+];
+export const colors: Color[] = [
+  "primary",
+  "onPrimary",
+  "primaryContainer",
+  "onPrimaryContainer",
+  "inversePrimary",
+  "secondary",
+  "onSecondary",
+  "secondaryContainer",
+  "onSecondaryContainer",
+  "tertiary",
+  "onTertiary",
+  "tertiaryContainer",
+  "onTertiaryContainer",
+  "error",
+  "onError",
+  "errorContainer",
+  "onErrorContainer",
+  "background",
+  "onBackground",
+  "surface",
+  "onSurface",
+  "surfaceVariant",
+  "onSurfaceVariant",
+  "inverseSurface",
+  "inverseOnSurface",
+  "outline",
+  "outlineVariant",
+  "shadow",
+  "scrim",
+  "surfaceDim",
+  "surfaceBright",
+  "surfaceContainerLowest",
+  "surfaceContainerLow",
+  "surfaceContainer",
+  "surfaceContainerHigh",
+  "surfaceContainerHighest",
+  "surfaceTint",
+];
+
 export const genCSS = (light: SerializedScheme, dark: SerializedScheme) => {
   const genColorVariable = (name: string, argb: number) => {
     const kebabCase = name.replace(/[A-Z]/g, (letter: string) => `-${letter.toLowerCase()}`);
@@ -73,4 +129,12 @@ ${darkColors}
   }
 }`;
   return colors;
+};
+
+export const serializeScheme = (scheme: DynamicScheme) => {
+  const out: Record<string, number> = {};
+  for (const color of colors) {
+    out[color] = MaterialDynamicColors[color as Color].getArgb(scheme);
+  }
+  return out as SerializedScheme;
 };
